Class Creation Progress
Added to the repo classes and Wrote a few basic JUnit tests for the Carrier and LookupType Repos

+++ Back-End/Springboot/Paychex-Dashboard/src/main/java/com/gcucapstone/paychexdashboard/entity/LookupType.java
@@ -35,7 +35,7 @@ public class LookupType {
 	//----------------------------------------------------
 	@Id
 	@Column(name = "lookup_type_id")
-	private int lookupTypeId;
+	private Long lookupTypeId;
 
 
 	@Column(name = "lookup_type")
